pyproject.toml
[project]
name = "py"
version = "0.1.0"
description = ""
readme = "readme.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.5",
    "logfire>=3.5.3",
    "mlx>=0.23.0",
    "mlx-vlm>=0.3.7",
    "mss>=9.0.1",
    "numpy>=2.2.5",
    "pillow>=11.1.0",
    "pydantic>=2.10.6",
    "pydantic-ai>=0.0.24",
    "python-dotenv>=1.0.0",
    "rich>=13.9.4",
    "safetensors>=0.5.2",
    "textual>=0.79.0",
    "torch>=2.6.0",
    "typer>=0.13.1",
    "uvicorn>=0.34.0",
]

[dependency-groups]
dev = ["ruff>=0.9.4"]

[project.scripts]
flow = "cli.flow.__main__:entrypoint"
py-scripts = "cli.scripts_cli.__main__:entrypoint"
snapvision = "cli.snapvision.__main__:main"
